+++ js/categories/categories.js
@@ -1,4 +1,5 @@
 import { categories } from '../../categories/categories.js';
+import {fetchFile} from "../utilities/utilities.js";
 
 export function showCategories(container) {
 
@@ -65,62 +66,10 @@ function addTemplates(container, category) {
 export function getTemplate(dir, files, callback) {
   
   files.split(",").forEach(f => {
-    getFile(`${dir}/${f.trim()}`, callback);
+     const file = `${dir}/${f.trim()}`;
+     fetchFile(file, function(rawFile) {
+        callback(file, rawFile);  
+     });
   });    
   
-}
-
-function getFile(file, callback) {
-
-  fetch(file)
-  .then(function (response) {
-    if (!response.ok) {
-      throw new Error(`HTTP error on ${file}. Status:${response.status}`);
-    }
-    return response.blob();
-  })
-  .then(function (myBlob) {
-    var reader = new FileReader();
-    var textFile = /text.*/;
-    if (myBlob.type.match(textFile)) {
-
-      reader.onload = function (event) {         
-        callback(file, event.target.result);       
-      };
-
-    } else {
-      console.log(`HTTP exception on ${file}`);
-    }
-    reader.readAsText(myBlob);
-  })
-  .catch(function (error) {
-    console.log(`HTTP exception on ${file}. ${error}`);
-  });
-}
-
-/* SEARCH */
-
-export function filterTemplates() {
-  
-  var filter, templates, div, item, i, txtValue;
-  
-  filter = this.value.toUpperCase();  
-  templates = document.getElementById("categoryTemplatesList");
-  div = templates.getElementsByTagName("div");
-
-  // Treats lists items like an array, where each item can be accessed through      it's index
-  for (i = 0; i < div.length; i++) {
-    item = div[i];
-
-    // All list item are checked to see if the value of the input, 
-    // ignoring case, matches the inner text or inner html of the item.
-    txtValue = item.textContent || item.innerText;
-
-    if (txtValue.toUpperCase().indexOf(filter) > -1) {
-      // Displays list items that are a match, and nothing if no match
-      div[i].style.display = "";
-    } else {
-      div[i].style.display = "none";
-    }
-  }
 }